virtio-blk: add nop-mode=on|off option for performance analysis
The new -device virtio-blk-pci,drive=foo,nop-mode=on option causes the
virtio-blk device to immediately complete all requests successfully.
The data buffers are not touched at all and it is common to see dirty
guest page contents when reading from the device.

This mode is useful for performance analysis when the aim is to study
the efficiency of virtio-blk or the guest kernel I/O code path.  Care
must be taken since zero-duration requests do not model real-world
request durations properly and may bypass batching or interrupt
mitigation optimizations.

Make sure to keep nop-mode off for the boot drive:

  $ qemu -drive if=none,id=root,file=rhel-6.img,cache=none \
         -device virtio-blk-pci,drive=root \
         -drive if=none,id=test,file=test.img,cache=none \
         -device virtio-blk-pci,drive=test,nop-mode=on

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>