qed: warn when in-use L2 tables are evicted
Cached L2 tables always have a refcount of 1.  When the cache is full
the first entry will be evicted.  It's unsafe to evict entries with
refcount > 1 because this can result in stale L2 tables due to
overlapping metadata I/O.

This patch is just for debugging, warn when such tables are evicted from
the cache.  This can be used together with overlapping I/O warnings to
show that we are indeed at risk of using stale L2 data.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>