net: convert NetClientState to QOM
This patch makes all net clients QOM objects.  This is the first step to
converting the net subsystem to the QEMU Object Model.  At this point net
clients are not yet visible in the QOM tree and there are many simplifications
that can still be made to drop ad-hoc code in favor of QOM.

The major changes are:

1. NetClientInfo becomes NetClientClass.  The "virtual functions" (aka
   callbacks) for net clients are defined here and overridden by
   NetClient child classes.

2. NetClientState becomes a QOM object.

3. NICState becomes a QOM object with NetClientState as its base class.
   In other words, a NIC is-a net client.

All net backends in net/ and NICs in hw/ must be converted to NetClient
subclasses.  This makes up the bulk of this patch and was done automatically
using a script.  Unfortunately there is no clean way to split these changes
into separate patches.

The net_client_type enum has been dropped.  For now users have been replaced
with object_dynamic_cast() to test the type of an object.  The cleaner solution
is to design polymorphic interfaces instead of sprinkling code that tests the
specific type of an object throughout the net subsystem.

Note that extrafx_eth.c uses tabs and checkpatch.pl will complain.  The rest of
the hunks are clean.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>